Order times will handle multiple units.
OrderXXXClamp uses nextOrderTime to get delay between orders. There is a
problem, for example, with Wildsoul. When hero is commanded Booboo needs
to way that delay. This fix creates own delay for each unit, so that in
one frame bot brain can command multiple units with delaying unit with
its own specific delay.